More misc fixes

+++ src/com/mraof/minestuck/world/lands/LandAspectRegistry.java
@@ -234,7 +234,7 @@ public class LandAspectRegistry
 		int id = SkaianetHandler.enterMedium((EntityPlayerMP)player, newLandId);
 		if(id != newLandId)	//Player already got a land, but the tag was somehow lost?
 			newLandId = id;
-		else MinestuckDimensionHandler.setSpawn(newLandId, new BlockPos(player.posX, 128 - MinestuckConfig.artifactRange, player.posZ));
+		else MinestuckDimensionHandler.setSpawn(newLandId, new BlockPos((int) player.posX, 128 - MinestuckConfig.artifactRange, (int) player.posZ));
 		
 		MinestuckPlayerTracker.updateLands();
